+++ changelog.md
@@ -39,6 +39,8 @@ This version is not yet released. If you are reading this on the website, then t
   - The inverse of the cosine idiom created a logical inconsistency
 - Flip the order of [`send`](https://uiua.org/docs/send)'s arguments
 - Add the [`pool`](https://uiua.org/docs/pool) modifier, which is identical to [`spawn`](https://uiua.org/docs/spawn) but spawns a thread in a thread pool
+- [`&rs`](https://uiua.org/docs/&rs) can now take a count of [`infinity ∞`](https://uiua.org/docs/infinity) to read until the end of the stream
+- Add the [`&runs`](https://uiua.org/docs/&runs) system function, which runs a command and returns an IO stream handle
 - Add the experimental [`&ffi`](https://uiua.org/docs/&ffi) system function, which allows calling functions from shared libraries
   - FFI is still a work in progress, but it is currently useful for foreign functions that aren't too complex
   - Bindings for [Raylib](https://www.raylib.com/) are being worked on as a proof of concept. You can find them in the [rayua](https://github.com/uiua-lang/rayua) repository.